prilozhenie/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.29
python-dotenv==1.0.1
jinja2==3.1.3
pandas==2.2.1
openpyxl==3.1.2
python-dateutil==2.8.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
aiofiles==24.1.0
tabulate==0.9.0